changing the return value as a decimal

+++ MySQL_Advanced/10-div.sql
@@ -2,10 +2,10 @@
 -- Script that creates a function that divides and returns something
 DELIMITER //
 CREATE FUNCTION SafeDiv(a INT, b INT)
-RETURNS INT
+RETURNS DECIMAL(10, 10)
 DETERMINISTIC
 BEGIN
-  DECLARE result FLOAT;
+  DECLARE result DECIMAL(10, 10);
   IF b = 0 THEN
     SET result = 0;
   ELSE
